cleanup and fixes

+++ src/C-DEngine/C-DEngines/C-DStorageService/cde_StorageMirrorCore.cs
@@ -1071,7 +1071,8 @@ namespace nsCDEngine.Engines.StorageService
 
                                 try
                                 {
-                                    System.IO.File.Move(fileToReturn, fileToReturn + ".1");
+                                    if (System.IO.File.Exists(fileToReturn))
+                                        System.IO.File.Move(fileToReturn, fileToReturn + ".1");
                                 }
                                 catch
                                 {
